[REFECTOR] redux store 로직 변경
