+++ src/store.js
@@ -1,0 +1,6 @@
+import { createStore } from "redux";
+import rootReducer from "./reducer";
+
+const store = createStore(rootReducer)
+
+export default store
